+++ lib/easytag/attributes/base.rb
@@ -1,0 +1,103 @@
+module EasyTag::Attributes
+  # for type casting
+  module Type
+    STRING      = 0
+    INT         = 1
+    FLOAT       = 2
+    INT_LIST    = 3
+    STRING_LIST = 4
+    BOOLEAN     = 5
+    DATETIME    = 6
+  end
+  class BaseAttribute
+    Utilities = EasyTag::Utilities
+
+    def initialize(args)
+      @name = args[:name]
+      @default = args[:default]
+      @type = args[:type] || Type::STRING
+      @options = args[:options] || {}
+      @ivar = BaseAttribute.name_to_ivar(@name)
+
+      if args[:handler].is_a?(Symbol)
+        @handler = method(args[:handler])
+      elsif args[:handler].is_a?(Proc)
+        @handler = args[:handler]
+      end
+
+      # fill default options
+      
+      # Remove nil objects from array (post process)
+      @options[:compact]    ||= false
+      # normalizes key (if hash) (handler)
+      @options[:normalize]  ||= false
+      # cast key (if hash) to symbol (handler)
+      @options[:to_sym]     ||= false
+
+    end
+    
+    def self.can_clone?(obj)
+      obj.is_a?(String) || obj.is_a?(Array) || obj.is_a?(Hash)
+    end
+
+    def self.deep_copy(obj)
+      Marshal.load(Marshal.dump(obj))
+    end
+
+    def default
+      BaseAttribute.can_clone?(@default) ?
+        BaseAttribute.deep_copy(@default) : @default
+    end
+    
+    def call(iface)
+      #puts 'entered call()'
+      data = @handler.call(iface)
+      data = type_cast(data)
+      post_process(data)
+    end
+
+    def type_cast(data)
+      case @type
+      when Type::INT
+        data = data.to_i
+      when Type::DATETIME
+        data = Utilities.get_datetime(data.to_s)
+      end
+      
+      data
+    end
+
+    def post_process(data)
+      if @options[:is_flag]
+        data = data.to_i == 1 ? true : false
+      end
+
+      # fall back to default if data is nil
+      data = BaseAttribute.obj_or_nil(data) || default
+
+      # run obj_or_nil on each item in array
+      data.map! { |item| BaseAttribute.obj_or_nil(item) } if data.is_a?(Array)
+
+      if @options[:compact] && data.respond_to?(:compact!)
+        data.compact!
+      end
+
+      data
+    end
+    # avoid returing empty objects
+    def self.obj_or_nil(o)
+      if o.class == String
+        ret = o.empty? ? nil : o
+      else
+        o
+      end
+    end
+
+    def self.name_to_ivar(name)
+      name = name.to_s if name.class == Symbol
+      name.gsub!(/\?/, '')
+      name.insert(0, '@')
+      name.to_sym
+    end
+  end
+end