+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-##### v0.2.0 (Not yet released) #####
+##### v0.3.0 (2013-05-29) #####
+* added:
+  - `#encoded_by`
+  - `#encoder_settings`
+  - `#group`
+  - `#composer`
+  - `#lyrics`
+  - `#compilation?`
+  - `#subtitle`
+  - `#bpm`
+  - `#lyricist`
+  - `#copyright`
+  - `#comment`
+
+* changed:
+  - `#comments` now returns an array, `#comment` is the
+    equivalent of the old behavior
+
+##### v0.2.0 (2013-05-25) #####
 * added:
   - `#track_num`
   - `#disc_num`